Add search results page for loop54
Carve out the results part of the search custom element and make that available on the search page.
