+++ content/search.md
@@ -1,0 +1,4 @@
+---
+layout: search
+noindex: true
+---
