+++ layouts/_default/baseof.html
@@ -42,6 +42,7 @@
   {{ $.Scratch.Add "css-base" "partials/elements/r-search.css" }}
   {{ else if site.Params.public.loop54 }}
   {{ $.Scratch.Add "css-base" "partials/elements/r-search-loop.css" }}
+  {{ $.Scratch.Add "css-base" "partials/elements/r-search-results.css" }}
   {{ end }}
   <!-- if benefits enabled, add that css and add ts here as well -->
   {{ if $announcements.benefitsEnabled }}
