+++ layouts/_default/search.css
@@ -1,0 +1,5 @@
+main {
+  padding: var(--block-padding);
+  margin: var(--block-margin);
+  max-width: var(--block-width);
+}
