+++ layouts/_default/search.html
@@ -1,0 +1,15 @@
+{{ define "dependencies" }}
+<meta name="robots" content="noindex">
+{{ $.Scratch.Add "css-critical" "_default/search.css" }}
+{{ $.Scratch.Add "css-critical" "partials/elements/r-search-results.css" }}
+{{ end }}
+
+{{ define "main" }}
+
+{{ $tsname := "_default/search.ts" }}
+{{- $js := (resources.Get $tsname) | js.Build (dict "format" "esm") -}}
+<script src="{{$js.RelPermalink}}" type="module"></script>
+
+{{partial "elements/r-search-results" (dict "take" 50)}}
+
+{{ end }}
