+++ layouts/partials/elements/r-search-results.html
@@ -1,0 +1,37 @@
+<r-search-results id="search-results" class="search-results"
+  loop-url="{{site.Params.public.loop54.url}}" currency="{{site.Params.public.currency}}" locale="{{site.Language.Lang}}"
+  {{ with .take }}take="{{.}}"{{ end }}
+>
+  <div results="suggested" show>
+    <h3 class="search-results__list-title">Suggested</h3>
+    <ul class="search-results__link-list">
+      <li><a href="" class="silent-link">Raingear</a></li>
+      <li><a href="" class="silent-link">Jackets</a></li>
+      <li><a href="" class="silent-link">Rubber Boots</a></li>
+      <li><a href="" class="silent-link">Outerwear</a></li>
+      <li><a href="" class="silent-link">Fleece jackets</a></li>
+      <li><a href="" class="silent-link">Sneakers</a></li>
+    </ul>
+  </div>
+  <div results="categories">
+    <ul class="search-results__link-list"></ul>
+  </div>
+  <div results="products">
+    <ul class="search-results__product-list"></ul>
+  </div>
+  <div results="more">
+    <a href="" class="btn">Show all</a>
+  </div>
+  <template product>
+    <li>
+      <a href="" class="silent-link">
+        <img src="" alt="">
+        <h4 title></h4>
+        <span price></span>
+      </a>
+    </li>
+  </template>
+  <template category>
+    <li><a href="" class="silent-link"></a></li>
+  </template>
+</r-search-results>
